% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chopthin}
\alias{chopthin}
\title{The Chopthin Resampler}
\usage{
chopthin(w, N, eta = 5.828427)
}
\arguments{
\item{w}{a vector of weights}

\item{N}{target number of particles}

\item{eta}{upper bound on the ratio between the weights. Must be >=4.}
}
\value{
A list with two elements: new weights and indices of the ancestors of the new particles. The weights are normalised to add up to N.
}
\description{
A CPP implementation of the Chopthin resampler. Can be used as the resampling step in particle filters and in sequential Monte Carlo.
}
\examples{
\dontrun{
chopthin(runif(10),N=10)
chopthin(runif(10),N=20,4)
chopthin(runif(10),N=5)
chopthin(runif(10),N=1)
}
}
