% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHS.R
\name{MHS}
\alias{MHS}
\title{A Beta distribution Metropolis-Hasting sampler (MHS) using R}
\usage{
MHS(n, alpha, beta, display = TRUE)
}
\arguments{
\item{n}{the number of samples}

\item{alpha}{Beta distribution alpha param}

\item{beta}{Beta distribution beta param}

\item{display}{whether to display a generated histogram}
}
\value{
a random sample of size \code{N}
}
\description{
This is a setup function, a Beta distribution Metropolis-Hasting sampler using R
}
\examples{
\dontrun{
    trials <- MHS(n=1e5, alpha=3, beta=2, display=TRUE)
}
}
