% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAM.R
\name{RAM}
\alias{RAM}
\title{A Robust Adaptive Metropolis (RAM) sampler using R}
\usage{
RAM(
  n,
  p,
  init,
  scale = rep(1, length(init)),
  log = TRUE,
  adapt = !is.null(acc.rate),
  acc.rate = NULL,
  gamma = 0.55,
  list = TRUE,
  n.start = 0
)
}
\arguments{
\item{n}{number of samples}

\item{p}{function that returns a value proportional to the log probability density to sample from.}

\item{init}{initial values vector}

\item{scale}{vector with the variances}

\item{log}{whether to sampled from a log density}

\item{adapt}{whether to use adaptive method}

\item{acc.rate}{desired acceptance rate}

\item{gamma}{controls the speed of adaption. Should be between 0.5 and 1. A lower gamma leads to faster adaption.}

\item{list}{whether to return a list (if False a matrix will be returned)}

\item{n.start}{iteration where the adaption starts. Only internally used.}
}
\value{
If `list` is TRUE a list is returned otherwise only a matrix with the samples
}
\description{
R implementation of RAM
}
\examples{
\dontrun{
  #log-pdf to sample from
  p.log <- function(x) {
    B <- 0.03
    -x[1]^2/200 - 1/2*(x[2]+B*x[1]^2-100*B)^2
  }
  # generate samples
  1. non-adaptive sampling
  samples.1 <- RAM(n=100, p.log, init=c(0,1), scale=c(1,0.1), adapt=FALSE)
  2. adaptive sampling
  samples.1 <- RAM(n=100, p.log, init=c(0,1), scale=c(1,0.1), adapt=TRUE, acc.rate=0.2)
}
}
