\name{SA23229030-package}
\alias{SA23229030-package}
\alias{SA23229030}
\docType{package}
\title{
  The final R package of 'Statistical Computing' course.
}
\description{
  Implementations of several samplers and resamplers.
}
\author{
Zhenjie Liu 

Maintainer: Zhenjie Liu <liuzhenjie@mail.ustc.edu.cn>
}

\keyword{ package }
